Add new abstract rule file option
This currently doesn't do anything but is in preparation for supporting
abstract Unix domain sockets. Only the rule (file) option is added here
with its various constraints.

The reason why I duplicated quite a bit in validate_rule is that I want
to actually have descriptive error messages and possibly I18N in the
future.

Signed-off-by: aszlig <aszlig@nix.build>