[WIP] Pass along abstract rule file option
The implementation here is still a bit weird and I really want to avoid
to introduce another boolean. Essentially what distinguishes a file
system socket from an abstract socket is the leading \0 and the
truncated length of struct sockaddr_un.

What this essentially does is just adding a bunch of ugly conditions and
essentially pass everything to Socket::format_sockpath in the hopes that
this function will one day do the right thing[TM].

A better option however would be to handle construction of struct
sockaddr_un in a dedicated function, since Socket::format_sockpath only
returns a std::string, which is not enough given that we either need to
take into account a terminating \0 (for file system sockets) or we don't
and directly truncate struct sockaddr_un at abstract.size() +
sizeof(sa_family_t).

Signed-off-by: aszlig <aszlig@nix.build>